kkin/requirements.txt
anyio==4.2.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==6.0.0
flake8-isort==6.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
isort==5.13.2
mccabe==0.7.0
psycopg2==2.9.9
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
sniffio==1.3.0
sqlparse==0.4.4
urllib3==2.1.0
